mejoras de codigo fuente

+++ src/main/java/classes/Personatge.java
@@ -45,7 +45,7 @@ public class Personatge {
         this.fuerza = fuerza;
     }
 
-    public double getVida() {
+    public int getVida() {
         return vida;
     }
